backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

sqlalchemy==2.0.29
mysql-connector-python==8.3.0

pydantic==2.6.4
pydantic-settings==2.2.1

python-dotenv==1.0.1

openai==1.14.3

alembic==1.13.1

httpx==0.27.0

passlib[bcrypt]==1.7.4
bcrypt==4.1.2